+++ crates/miden-testing/src/tx_context/builder.rs
@@ -676,11 +676,11 @@ impl TransactionContextBuilder {
 
         let mast_store = {
             let mast_forest_store = TransactionMastStore::new();
-            mast_forest_store.load_transaction_code(
-                tx_inputs.account().code(),
-                tx_inputs.input_notes(),
-                &tx_args,
-            );
+            mast_forest_store.load_account_code(tx_inputs.account().code());
+
+            for acc_inputs in tx_args.foreign_account_inputs() {
+                mast_forest_store.insert(acc_inputs.code().mast());
+            }
 
             mast_forest_store
         };