+++ crates/miden-tx/src/prover/mast_store.rs
@@ -1,12 +1,7 @@
 use alloc::{collections::BTreeMap, sync::Arc};
 
 use miden_lib::{MidenLib, StdLibrary, transaction::TransactionKernel, utils::sync::RwLock};
-use miden_objects::{
-    Digest,
-    account::AccountCode,
-    assembly::mast::MastForest,
-    transaction::{InputNote, InputNotes, TransactionArgs},
-};
+use miden_objects::{Digest, account::AccountCode, assembly::mast::MastForest};
 use vm_processor::MastForestStore;
 
 // TRANSACTION MAST STORE
@@ -50,38 +45,6 @@ impl TransactionMastStore {
         store
     }
 
-    /// Loads code required for executing a transaction with the specified inputs and args into
-    /// this store.
-    ///
-    /// The loaded code includes:
-    /// - Account code for the account specified from the provided [AccountCode].
-    /// - Note scripts for all input notes in the provided [InputNotes].
-    /// - Transaction script (if any) from the specified [TransactionArgs].
-    pub fn load_transaction_code(
-        &self,
-        account_code: &AccountCode,
-        input_notes: &InputNotes<InputNote>,
-        tx_args: &TransactionArgs,
-    ) {
-        // load account code
-        self.load_account_code(account_code);
-
-        // load note script MAST into the MAST store
-        for note in input_notes {
-            self.insert(note.note().script().mast().clone());
-        }
-
-        // add extra account codes
-        for foreign_account in tx_args.foreign_account_inputs() {
-            self.load_account_code(foreign_account.code());
-        }
-
-        // load tx script MAST into the MAST store
-        if let Some(tx_script) = tx_args.tx_script() {
-            self.insert(tx_script.mast().clone());
-        }
-    }
-
     /// Registers all procedures of the provided [MastForest] with this store.
     pub fn insert(&self, mast_forest: Arc<MastForest>) {
         let mut mast_forests = self.mast_forests.write();
@@ -93,7 +56,7 @@ impl TransactionMastStore {
     }
 
     /// Loads the provided account code into this store.
-    fn load_account_code(&self, code: &AccountCode) {
+    pub fn load_account_code(&self, code: &AccountCode) {
         self.insert(code.mast().clone());
     }
 }